- Moved service registration at infrastructure - Created class for IdentityConfiguration
